promicroish: add SuperMini nRF52840
I created a board w/ the nice!nano footprint but, after
receiving the board, noticed that the footprint was
slightly different than expected.

Using the exported footprint from the wiki page, I
created the same set of footprints the nice!nano has
and a symbol (exactly the same as the nice!nano)

I also changed two of the pin names for the nice!nano symbol
as they had a "-" instead of a "." in the "Px.xx" pin name.

Signed-off-by: Helmut Lord <kellyhlord@gmail.com>